+++ src/main/java/gtclassic/common/GTBlocks.java
@@ -155,6 +155,7 @@ public class GTBlocks {
 	public static final GTBlockMachine tileTranslocatorFluid = registerBlock(new GTBlockMachine("translocatorfluid", GTLang.TRANSLOCATOR_FLUID, 1));
 	public static final GTBlockMachine tileBufferFluid = registerBlock(new GTBlockMachine("bufferfluid", GTLang.BUFFER_FLUID, 1));
 	public static final GTBlockLamp tileLamp = registerBlock(new GTBlockLamp());
+	public static final GTBlockSuperconductorCable tileSuperconductorCableBEACON = registerBlock(new GTBlockSuperconductorCable(16, "0"));
 	public static final GTBlockSuperconductorCable tileSuperconductorCableMAX = registerBlock(new GTBlockSuperconductorCable(12, ""));
 	public static final GTBlockSuperconductorCable tileSuperconductorCableIV = registerBlock(new GTBlockSuperconductorCable(6, "2"));
 	public static final GTBlockSuperconductorCable tileSuperconductorCableHV = registerBlock(new GTBlockSuperconductorCable(4, "4"));